added sqlite support and some other refactoring
