+++ src/server.c
@@ -1,4 +1,5 @@
 #include "server.h" // https://man7.org/linux/man-pages/man2/listen.2.html
+// #include "helpers/database_operations.h"
 
 int main(void) {
     int master_socket, new_connected_fd; // File descriptor and new socket created on acceptance
@@ -47,7 +48,6 @@ int main(void) {
     int maxfd;
     int sd = 0;
     int fds_to_read;
-    ssize_t bytes_sent;
 
     while (true) {
         FD_ZERO(&readfds);
@@ -113,6 +113,7 @@ int main(void) {
                     print_stdout("message from client\n");
                     print_stdout(buffer);
                     context = InitContext();
+                    // ExecuteQuery("TIMBOG_DEV", "name=snorlax", context->message_bus);
                     HandleClientRequest(sd, buffer, context);
                     free(context);
                     close(sd);
