+++ src/shared.h
